placing Bets backend

+++ sep/src/main/java/com/gruppe_f/sep/entities/leagueData/LeagueDataController.java
@@ -21,7 +21,11 @@ public class LeagueDataController {
     private final DateRepository repo2;
 
     @Autowired
-    public LeagueDataController(LeagueDataRepository repo, DateRepository repo2) {this.repo = repo; this.repo2= repo2;}
+    public LeagueDataController(LeagueDataRepository repo, DateRepository repo2) {
+        this.repo = repo;
+        this.repo2 = repo2;
+    }
+
     @PostMapping("/add")
     public ResponseEntity<LeagueData> addLeagueData(@RequestBody LeagueData data) {
         LeagueData newData = repo.save(data);
@@ -40,10 +44,10 @@ public class LeagueDataController {
         List<LeagueData> list = repo.findAll();
         List<SystemDate> sysDate = repo2.findAll();
         List<LeagueData> returnList = new ArrayList<>();
-        for(LeagueData data:list) {
+        for (LeagueData data : list) {
             if (data.getLiga().getId() == id) {
                 //Getting LeagueData by ID and setting result of Future games "0-0"
-                if(data.getDate().compareTo(sysDate.get(0).getLocalDate()) < 0) {
+                if (data.getDate().compareTo(sysDate.get(0).getLocalDate()) < 0) {
                     data.getLiga().setLeagueData(null);
                     returnList.add(data);
                 } else {
@@ -54,20 +58,21 @@ public class LeagueDataController {
             }
         }
         //Sorting Resulting LeagueData List by Date
-        List<LeagueData> result = returnList.stream().sorted((x, y)-> x.getDate().compareTo(y.getDate())).collect(Collectors.toList());
+        List<LeagueData> result = returnList.stream().sorted((x, y) -> x.getDate().compareTo(y.getDate())).collect(Collectors.toList());
 
         return new ResponseEntity<>(result, HttpStatus.OK);
 
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<LeagueData> getData(@PathVariable("id") int id) {
+    public ResponseEntity<LeagueData> getData(@PathVariable("id") Long id) {
         LeagueData data = repo.findByid(id);
         data.getLiga().setLeagueData(null);
         return new ResponseEntity<>(data, HttpStatus.OK);
     }
+
     @PutMapping("/update/{id}")
-    public ResponseEntity<?> updateLeagueData(@PathVariable("id") int id, @RequestBody LeagueData data) {
+    public ResponseEntity<?> updateLeagueData(@PathVariable("id") Long id, @RequestBody LeagueData data) {
         LeagueData newData = repo.findByid(id);
         newData.getLiga().setLeagueData(null);
         newData.setResult(data.getResult());
@@ -77,4 +82,5 @@ public class LeagueDataController {
         newData.setMatchDay(data.getMatchDay());
         return new ResponseEntity<>(repo.save(newData), HttpStatus.OK);
     }
+
 }
